Refactor Local & Url CSV reader as well as tests, fix transaction per Zander's suggestion
